Refactor form service code into various libs
